+++ src/engine/mod.rs
@@ -1,7 +1,6 @@
 mod evaluation;
 mod field;
 mod function;
-mod gradients;
 mod literal;
 mod network;
 mod symbol;
@@ -9,8 +8,7 @@ mod tape;
 mod value;
 
 pub use evaluation::Evaluation;
-pub use field::Field;
-pub use gradients::Gradients;
+pub use field::{Field, Gradients};
 pub use network::Network;
 pub use symbol::Symbol;
 pub use value::Value;